+++ bindings/nodejs/types/client/index.ts
@@ -8,7 +8,7 @@ export * from './ledgerNanoStatus';
 export * from './network';
 export * from './nodeInfo';
 export * from './outputIdsResponse';
-export * from './outputBuilderOptions';
+export * from './outputBuilderParams';
 export * from './preparedTransactionData';
 export * from './queryParameters';
 export * from './range';
