+++ bindings/nodejs/types/client/outputBuilderParams/nftOutputParams.ts
